wip: refactor vesting types
Now that `serde_yaml` is out of the picture,
`#[serde(tag=...)]` and not being able to nest structs
are also a thing of the past.

So the shared fields for all vesting types (`amount` and `addr`)
are kept at the top level but a new `Vesting` struct is introduced
(which still has some repetition).

This saves some weirdness around `match`ing the vesting type.