chart: render updated schedule
This finally replaces the ugly JS+PUG that rendered `docs/chart.svg`
with some nice safe Rust to generate a more detailed version of the
schedule (based on `mgmt/src/schedule.rs`) at `docs/schedule.svg`.
Run `make schedule` to produce it.

The old `docs/chart.svg` is removed; the rest of the `docs/` scripts
will be removed as soon as a new transaction list is generated.