refactor: extract tests while looking for bug
spoiler: it was a missing `?` mark. it was serializing
`{"Ok":{"foo"...}}` instead of `{"foo"...` and then
expecting `{"foo"...` in the deserialized object.

this broke half the tests. since the opaque mocked storage
was in the way of the control flow, printing the serialized
state as it was being saved/loaded was what identified the
extra level of nesting